renamed UserPreferencesController to AddGenreScene and removed old function. made obj-c function work , adapted storyboard to work with it.
